requirements.txt
fastapi==0.75.1
lightgbm==3.3.2
numpy==1.20.3
pandas==1.3.4
pydantic==1.9.0
uvicorn==0.15.0
PyYAML==6.0
scikit_learn==1.0.2
